Add example plot to document.
Note that the actual plots are ignored by Git, because they change with
every run of the notebook.

It *is* possible to make a custom rule in the Makefile for latexmk to
build each of the plots by executing the notebook with
    jupyter-nbconvert --execute --inplace prob1.ipynb
but this takes far more time than I'm willing to wait to build a PDF.
Although it should be possible to get make to run the notebooks only if
the files are missing...

This will have to be something to look into in the future. For now, just
run the notebook as many times as it takes to get good looking plots.